requirements.txt
torch
ultralytics
opencv-python
Pillow
numpy
arabic-reshaper
python-bidi
tk
